Generate gibberish from English CSV; use non-deprecated option to initialize file watcher for dev server
